+++ src/baderkit/core/methods/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+
+from .neargrid import NeargridMethod
+from .ongrid import OngridMethod
+from .pseudo_neargrid import PseudoNeargridMethod
+from .weight import WeightMethod
