+++ src/baderkit/core/methods/neargrid/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+from .neargrid_method import NeargridMethod
