+++ src/baderkit/core/methods/ongrid/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+from .ongrid_method import OngridMethod
