+++ src/baderkit/core/methods/ongrid/ongrid_numba.py
@@ -1,0 +1,66 @@
+# -*- coding: utf-8 -*-
+
+import numpy as np
+from numba import njit, prange
+from numpy.typing import NDArray
+
+from baderkit.core.methods.shared_numba import get_best_neighbor
+
+
+@njit(parallel=True, cache=True)
+def get_steepest_pointers(
+    data: NDArray[np.float64],
+    initial_labels: NDArray[np.int64],
+    neighbor_transforms: NDArray[np.int64],
+    neighbor_dists: NDArray[np.int64],
+    vacuum_mask: NDArray[np.bool_],
+):
+    """
+    For each voxel in a 3D grid of data, finds the index of the neighboring voxel with
+    the highest value, weighted by distance.
+
+    Parameters
+    ----------
+    data : NDArray[np.float64]
+        A 3D grid of values for each point.
+    initial_labels : NDArray[np.int64]
+        A 3D array where each entry represents the basin label of the point.
+    neighbor_transforms : NDArray[np.int64]
+        The transformations from each voxel to its neighbors.
+    neighbor_dists : NDArray[np.int64]
+        The distance to each neighboring voxel
+    vacuum_mask : NDArray[np.bool_]
+        A 3D array representing the location of the vacuum
+
+    Returns
+    -------
+    pointers : NDArray[np.int64]
+        A 3D array where each entry is the index of the neighbor that had the
+        greatest increase in value. A value of -1 indicates a vacuum point.
+
+    """
+    nx, ny, nz = data.shape
+    # create array to store the label of the neighboring voxel with the greatest
+    # elf value
+    pointers = initial_labels.copy()
+    # loop over each voxel in parallel
+    for i in prange(nx):
+        for j in range(ny):
+            for k in range(nz):
+                # check if this is a vacuum point. If so, we don't even bother
+                # with the label.
+                if vacuum_mask[i, j, k]:
+                    pointers[i, j, k] = -1
+                    continue
+                # get the best neighbor
+                best_transform, best_neigh, _ = get_best_neighbor(
+                    data=data,
+                    i=i,
+                    j=j,
+                    k=k,
+                    neighbor_transforms=neighbor_transforms,
+                    neighbor_dists=neighbor_dists,
+                )
+                x, y, z = best_neigh
+                pointers[i, j, k] = initial_labels[x, y, z]
+    return pointers
