+++ src/baderkit/core/methods/pseudo_neargrid/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+from .pseudo_neargrid_method import PseudoNeargridMethod
