+++ src/baderkit/core/methods/weight/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+from .weight_method import WeightMethod
